Convert unix versions to use our modified version of tcl/tk from the cvs. The advantage of this is that it has all the initialisation code built-in to the binaries and doesn't depend on any external libraries.
This means that you need to checkout and build the modules "tcl-8.0" and
"tk-8.0" from our cvs before you can build sdr (Note: compile, but don't
install these libraries. They don't conflict with the standard tcl/tk libs).

Yes, this is a bit of a pain, but it should stop us getting complaints
about tcl/tk interactions! We've used this for rat/nte for a while now
and it does seem to work....

Colin