Handle compressed announcements if zlib is available.  FreeBSD has zlib in the base system, so make FreeBSD compile this code by default. For other systems, CFLAGS has to have -DHAVE_ZLIB and LIBS has to have -L/wherever/zlib/is -lz .
Handle zlib-compressed (RFC1950) as well as gzip-compressed (RFC1952)
announcements, since zlib compression is easier with zlib and is
also the format the RealServer G2 uses (the only source of compressed
announcements that I know about).  Since zlib compression is easier
with the most commonly available compression library, it might make
sense to specify the use of zlib in SAP instead of gzip (especially
since gzip's extra header information buys SAP absolutely nothing.)

The SAP spec is not very clear on ordering of compression and
authentication; this patch performs authentication before decompression
(i.e. the authentication data has to authenticate the compressed
version of the announcement).  However, it probably makes sense to
compress after authentication (i.e. authenticate the cleartext
payload and then compress the authentication info along with the
payload).  For now, nobody's sending compressed authenticated SAP
so this isn't an issue.  The SAP spec should describe the desired
ordering and when it does so this code should be updated to follow it.